Poprawiłam funkcję generate_bipartite_graph

+++ Algoritms/graph_generators.py
@@ -14,7 +14,7 @@ def generate_cycle_graph(n):
 
 
 def generate_bipartite_graph(n):
-    return nx_graph_to_graph(nx.complete_bipartite_graph(n/2, n/2))
+    return nx_graph_to_graph(nx.complete_bipartite_graph(int(n/2), int(n/2)))
 
 
 def nx_graph_to_graph(nx_graph):  # Zamienia graf z biblioteki nx na słownik wierzchołków i krawędzi
